+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     * `--entrypoint` is now required if no inputs are provided.
     * `--entrypoint` will be prefixed to the key of any key-value pairs supplied on the command line.
 
+### Fixed
+
+* Output directory logic for `run` now properly reflects the documentation ([#147](https://github.com/stjude-rust-labs/sprocket/pull/147)).
+
 ## 0.14.1 - 07-10-2025
 
 ### Fixed